requirements.txt
guilded.py==1.9.1
motor==3.3.0
pymongo==4.5.0
toml~=0.10.2
roblox==2.0.0
aiohttp~=3.8.5
cachetools~=5.3.1